+++ web/js/utilities.js
@@ -162,4 +162,31 @@ export default {
             .catch(err => {return err})
     }
 }
+
+export function initializeTools(config) {
+    return new Promise((res) => {
+        const setContainer = document.getElementById("tool_set")
+        Array.from(config.REGISTRY).forEach(thumbnailGenerator => {
+            setContainer.innerHTML += thumbnailGenerator()
+        })
+        UTILS.broadcast(undefined, PLAYGROUND.EVENTS.LOADED, setContainer, {})
+        /**
+         * Really each render should be a promise and we should return a Promise.all() here of some kind.
+         * That would only work if PlaygroundRender resulted in a Promise where we could return Promise.all(renderPromises).
+         */
+        setTimeout(res, 200) //A small hack to ensure all the HTML generated by processing the views enters the DOM before this says it has resolved.
+    })
+}
+
+/**
+ * 
+ * @param {type} elem
+ * @param {type} viewURL
+ * @param {type} data
+ * @returns {undefined}
+ */
+export function initializeView(elem, viewURL, data) {
+    //fetch view HTML from URL
+    //Do playground magic on it (apply header, footer, and utilities)
+}
